Add test case for fulltext search
- Should find Tocco person in the fulltext search field in the app bar
- Opens the person
- Checks if the (readonly) detail form has been loaded and the
  firstname is "Support" and the lastname is "Tocco AG"
- Checks if the default display "Tocco AG, Support" is correctly
  displayed in the panel with the open entities
